// apps/web/src/api/generated/CatalogListDto.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { HfSearchResultDto } from "./HfSearchResultDto";
import type { InstalledModelDto } from "./InstalledModelDto";

export type CatalogListDto = { installed: Array<InstalledModelDto>, hf_results: Array<HfSearchResultDto>, hf_status: string, };