// apps/web/src/api/generated/ModelInstallTaskDto.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.

export type ModelInstallTaskDto = { task_id: string, model_install_id: string | null, routed_backend: string | null, routing_signal: string | null, progress_channel: string, cancel_endpoint: string, };